% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjoint_plot_bar.R
\name{conjoint_plot_bar}
\alias{conjoint_plot_bar}
\title{conjoint_plot_bar}
\usage{
conjoint_plot_bar(.x)
}
\arguments{
\item{.x}{A list or atomic vector.}
}
\value{
gggplot2 graph

a ggplot2 object
}
\description{
Plot results per dimension - Average Marginal Component Effects (AMCEs)
}
\examples{

kobodata <-  system.file("data-demo/conjoint_data.xlsx", package = "VulnerabilityScoreCalibration")
koboform <-  system.file("data-demo/conjoint_form.xlsx", package = "VulnerabilityScoreCalibration") 

cj <- conjoint_review(kobodata, koboform)
 
## Plot AMCES as bar for dimension 2
conjoint_plot_bar( as.data.frame(cj[["cjdata"]][2,][["amces"]])) + 
  ggplot2::labs( subtitle = "Average Marginal Component Effects (AMCEs)")

## Plot importance as bar for dimension 2
conjoint_plot_bar( as.data.frame(cj[["cjdata"]][2,][["importance"]])) + 
  ggplot2::labs( subtitle = "Importance")
}
