% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjoint_plot_point.R
\name{conjoint_plot_point}
\alias{conjoint_plot_point}
\title{conjoint_plot_point}
\usage{
conjoint_plot_point(.x)
}
\arguments{
\item{.x}{A list or atomic vector.}
}
\value{
a ggplot2 object
}
\description{
Plot results per dimension - Marginal Means
}
\examples{

kobodata <-  system.file("data-demo/conjoint_data.xlsx", package = "VulnerabilityScoreCalibration")
koboform <-  system.file("data-demo/conjoint_form.xlsx", package = "VulnerabilityScoreCalibration") 

cj <- conjoint_review(kobodata, koboform)
 
conjoint_plot_point( as.data.frame(cj[["cjdata"]][1,][["margins"]])) + 
  ggplot2::labs( subtitle = "Margins)")

conjoint_plot_point( as.data.frame(cj[["cjdata"]][1,][["amces"]])) + 
  ggplot2::labs( subtitle = "Average Marginal Component Effects (AMCEs)")

conjoint_plot_point( as.data.frame(cj[["cjdata"]][1,][["importance"]])) + 
  ggplot2::labs( subtitle = "Importance")
}
