% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjoint_review.R
\name{conjoint_review}
\alias{conjoint_review}
\title{conjoint_review}
\usage{
conjoint_review(kobodata, koboform)
}
\value{
a series of plot
}
\description{
What is Conjoint analysis?
 
 Conjoint analysis can speed up expert consultations by offering an 
 __objective mean to compile expert opinions__.
 
 * Conjoint analysis originated in mathematical psychology by psychometricians.
 
 * often used to evaluate how people make decisions between a set of 
 different options when considering a number of criteria at the same time 
 (conjoint features; “trade-offs”). 
 
 1. Measurement framework
 
 The [Joint Intersectoral Analysis Framework (JIAF)](https://interagencystandingcommittee.org/system/files/gbna_tools_and_guidance_pkg_final_june_2019.pdf) 
 is a theoretical generic measurement framework to be used for Humanitarian 
 needs assessment. It specifies three distinct and complementary components
  of humanitarian severity and vulnerability indexes:
   
 * Basic Needs & Living standards  
 
 * Coping Capacity
 
 * Well Being & Community integration
 
 This generic model can be contextualised: different sub-indicators might be 
 used for each of the 3 components depending on cultural and political situations. 
 
 2. Define the combined alternatives to be compared 
 
 * participants rate their preferences for profiles with different combinations
  of the attributes or criteria. 
 
 * CA then allows to “decompose” or reverse-engineer these ratings into
  estimates of how important each criteria or attribute is to a participant’s ranking decisions
 
 3. Utility scales & Agreement levels
 
 Estimating the contribution of each potential answers
 
 * Utility values indicate the overall contribution of each attribute to how 
 the profiles were rated (e.g. whether number of meals is more important in 
 vulnerability scoring than access to safe water). 
 
 * A higher _"utility"_ estimate indicates that this level contributes to a
  higher vulnerability than the level with the lower utility estimate 
  (it does not give an absolute value for the utility of an option, but rather
   assumes a reference alternative). 
 
 * Standard deviation for each level within model allows to better understand 
 how homogeneous the group of experts is with respect to one level.  
 
 4. Importance of each criteria 
 
  * Importance of each criteria represent the average importance as estimated from all experts.
 
  * Importance values will then be used as the weights for each attribute inside each of our three dimensions. 
 
  * Importance values sum to 100%.
}
\examples{

kobodata <-  system.file("data-demo/conjoint_data.xlsx", package = "VulnerabilityScoreCalibration")
koboform <-  system.file("data-demo/conjoint_form.xlsx", package = "VulnerabilityScoreCalibration") 

cj <- conjoint_review(kobodata, koboform)

cj[["data_quality"]]
}
