% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjoint_walk.R
\name{conjoint_walk}
\alias{conjoint_walk}
\title{conjoint_walk}
\usage{
conjoint_walk(dim, margins, amces, importance, ...)
}
\arguments{
\item{importance}{}
}
\description{
Print out a summary from the conjoint analysis
}
\details{
*  Marginal Means

  *  Average Marginal Component Effects (AMCEs)

  *  Average Marginal Component Effects (AMCEs)

  *  Importance Weights
}
\examples{
kobodata <-  system.file("data-demo/conjoint_data.xlsx", package = "VulnerabilityScoreCalibration")
koboform <-  system.file("data-demo/conjoint_form.xlsx", package = "VulnerabilityScoreCalibration") 

cj <- conjoint_review(kobodata, koboform) 
cjdata <- cj[["cjdata"]]
## Get a summary of all dimensions
purrr::pwalk(cjdata, conjoint_walk)

## Save a csv extract of the weights
# purrr::walk2(cjdata$dim, cjdata$amces, ~write_csv(.y, fs::path(.x, ext = "csv")))

all <- purrr::walk2(cjdata$amces,  ~cbind())

all <-purrr::pwalk(cjdata$amces, rbind)
all2 <- bind_rows(cjdata$amces, .id = "column_label")



}
