% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadratic_prepare.R
\name{quadratic_prepare}
\alias{quadratic_prepare}
\title{quadratic_prepare}
\usage{
quadratic_prepare()
}
\arguments{
\item{indicator}{a dataframe with max 5 groups of 5 indicators}
}
\description{
This function aims at quickly building a quadratic voting questionnaire
}
\details{
The questionnaire comes with limitation as it can process not more than 5 groups
of maximum 5 indicators. Each indicators is associated to different levels that
will be then assessed through conjoint analysis for the weighting stage

After quadratic voting, it is expected that the facilitation of the result 
interpretation should allow to reduce the numbers of indicators to a maximum of
12 indicators
}
\examples{

indicator <-  system.file("data-demo/indicator_criteria.xlsx", package = "VulnerabilityScoreCalibration")
quadratic_prepare()
}
