% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadratic_report.R
\name{quadratic_report}
\alias{quadratic_report}
\title{Generate Report From Quadratic Voting}
\usage{
quadratic_report(data = "data.xlsx", form = "form.xlsx", folder = "data-raw")
}
\arguments{
\item{data}{name of the file with data collected through kobotoolbox should be xlsx}

\item{form}{name of the file with form used to collected through kobotoolbox
quadratic survey (should be xlsx)}

\item{folder}{folder within your project where to put the generated report. 
Folder will be created if it does not exist}
}
\value{
nothing the file for the report is generated
}
\description{
Generate Report From Quadratic Voting
}
\examples{

## Example -> the report will be saved in the same folder...
# quadratic_report(data = "data.xlsx",
#                  form = "form.xlsx",
#                  folder = "data-raw")
}
