% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadratic_review.R
\name{quadratic_review}
\alias{quadratic_review}
\title{quadratic_review}
\usage{
quadratic_review(kobodata, koboform)
}
\arguments{
\item{kobodata}{path to data collected through kobotoolbox}

\item{koboform}{form used to collected through kobotoolbox quadratic survey}

\item{caption}{A reference to the excercise --}
}
\value{
list with data and standard plots
}
\description{
Explore the results from a quadratic voting consultations. 
 1. What are the prioritized Topics?
 2. How dispersed participants votes are? 
 3. Who is expecting or pushing back... on what?
}
\examples{

# kobodata <- here::here("", "quadra_data.xlsx")
# koboform <- here::here("", "survey_quadraticvoting_CBI_Indicators.xlsx")

kobodata <-  system.file("data-demo/quadra_data.xlsx", package = "VulnerabilityScoreCalibration")
koboform <-  system.file("data-demo/quadra_form.xlsx", package = "VulnerabilityScoreCalibration")
 
## Run the process
result <- quadratic_review(kobodata, koboform)

## Review output
result[["topic_prioritisation"]]

result[["vote_dispersion"]]

result[["individual_prioritisation"]]

}
