% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_quadratic_consultation.R
\name{review_quadratic_consultation}
\alias{review_quadratic_consultation}
\title{Title}
\usage{
review_quadratic_consultation()
}
\description{
Description
}
\examples{
review_quadratic_consultation()
}
