% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjoint_prepare.R
\name{conjoint_prepare}
\alias{conjoint_prepare}
\title{conjoint_prepare}
\usage{
conjoint_prepare(opts, language, form_title, id_string, outdir, outfile)
}
\arguments{
\item{opts}{a dataframe containing the options to compare}

\item{form_title}{Actividad #2: Calificación de perfiles de vulnerabilidad}

\item{id_string}{vulnerability_rating}

\item{outdir}{}

\item{outfile}{}
}
\value{
a file in xlsform format
}
\description{
Generate an xlsform used to perform consultation for conjoint analysis.
}
\details{
The options that will be compared are the different levels of the section 
of indicators that should have been filtered through the quadratic voting 
stage.

One level can actually match multiple response options from the screening 
questionnaire
}
\examples{


# indicator <-  system.file("data-demo/indicator_criteria.xlsx", package = "VulnerabilityScoreCalibration")
# opts <- read_excel("cja_opts_SAL.xlsx")
# 
# conjoint_prepare( opts = opts, 
#                   language = "Spanish (es)",
#                   form_title = "Actividad #2: Calificación de perfiles de vulnerabilidad - El Salvador, 23 de marzo, 2023",
#                   id_string = "vulnerability_rating",
#                   outdir = "", 
#                   outfile = "form.xlsx" )
}
